+++ src/main/java/wdmsystem/order/summary/ProductSummary.java
@@ -1,8 +1,10 @@
 package wdmsystem.order.summary;
 
+import java.math.BigDecimal;
+
 public record ProductSummary(
         String title,
-        double price,
+        BigDecimal price,
         TaxSummary tax,
         float weight,
         String weightUnit,
