+++ src/main/java/wdmsystem/order/summary/ProductVariantSummary.java
@@ -1,7 +1,9 @@
 package wdmsystem.order.summary;
 
+import java.math.BigDecimal;
+
 public record ProductVariantSummary(
         String title,
-        double additionalPrice
+        BigDecimal additionalPrice
 ) {
 }
